improved type safety, refactor AnyM into TraversableM and AnyM
