+++ cyclops-base/src/main/java/com/aol/cyclops/lambda/monads/Monads.java
@@ -159,10 +159,10 @@ public class Monads extends AsAnyM{
 	 * @param fn Function to apply 
 	 * @return Monad with a list
 	 */
-	public static <MONAD,R> Monad<MONAD,List<R>> traverse(Class c,List<?> seq, Function<?,R> fn){
-		return (Monad)asMonad(new ComprehenderSelector().selectComprehender(c).of(1))
+	public static <T,R> AnyM<List<R>> traverse(Class<?> c,List<T> seq, Function<?,R> fn){
+		return asMonad(new ComprehenderSelector().selectComprehender(c).of(1))
 								.flatMap(in-> asMonad(seq.stream()).flatMap(m-> m).flatMap((Function)fn).unwrap()
-									);
+									).anyM();
 	}
 
 	
@@ -182,9 +182,9 @@ public class Monads extends AsAnyM{
 	 * @param seq List of monads to convert
 	 * @return Monad with a List
 	 */ 
-	public static <MONAD,T>  Monad<MONAD,T> sequence(Class c, List<?> seq){
-		return (Monad)AsGenericMonad.asMonad(new ComprehenderSelector().selectComprehender(c).of(1))
+	public static <T,T1>  AnyM<List<T>> sequence(Class<?> c, List<T1> seq){
+		return AsGenericMonad.asMonad(new ComprehenderSelector().selectComprehender(c).of(1))
 				.flatMap(in-> asMonad(seq.stream()).flatMap(m-> m).unwrap()
-							);
+							).anyM();
 	}
 }
